Separate functionality on index.js and todoFunctions.js
